Compiling with DOLFINX 71e3e3fc9caef86917264f1efc3767e464f453d7 works. Should use `std::floating_point U = dolfinx::scalar_value_type_t<T>` wherever we have `PETScScalar` for easier deduction
