+++ cpp/CMakeLists.txt
@@ -67,13 +67,9 @@ install(FILES dolfinx_mpc.h  DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/dolfinx_mpc
 install(FILES assemble_utils.h mpi_utils.h ContactConstraint.h utils.h MultiPointConstraint.h SlipConstraint.h PeriodicConstraint.h assemble_matrix.h assemble_vector.h lifting.h mpc_helpers.h DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/dolfinx_mpc COMPONENT Development)
 # Add source files to the target
 target_sources(dolfinx_mpc PRIVATE
-${CMAKE_CURRENT_SOURCE_DIR}/SlipConstraint.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/ContactConstraint.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/PeriodicConstraint.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/utils.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/assemble_matrix.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/assemble_vector.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/mpc_helpers.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/assemble_utils.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/mpi_utils.cpp
   )
